ATA: Extract associated-token-account-client crate from associated-token-account crate (#7005)
* extract associated-token-address crate

* use spl_associated_token_address in spl crates

* use spl-program-ids in associated-token-address crate

* update imports in repo

* use solana-inline-spl instead of spl-program-ids

* remove remaining references to spl-program-ids

* Rename associated-token-address -> associated-token-account-client

* Use the new package everywhere

* Remove solana-inline-spl dependency

* Remove dependency on ata program from token-client

* Add #[doc(hidden)] attribute to internal function

* Remove ATA from token upgrade CLI

* Add program module for program id

---------
