+++ interface/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "spl-associated-token-account-client"
+version = "1.0.0"
+description = "Solana Program Library Associated Token Account Client"
+authors = ["Solana Labs Maintainers <maintainers@solanalabs.com>"]
+repository = "https://github.com/solana-labs/solana-program-library"
+license = "Apache-2.0"
+edition = "2021"
+
+[dependencies]
+solana-program = "2.0.3"
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
