+++ program-test/Cargo.toml
@@ -15,5 +15,6 @@ solana-program = "2.0.3"
 solana-program-test = "2.0.3"
 solana-sdk = "2.0.3"
 spl-associated-token-account = { version = "4.0.0", path = "../program", features = ["no-entrypoint"] }
+spl-associated-token-account-client = { version = "1.0.0", path = "../client" }
 spl-token = { version = "6.0", path = "../../token/program", features = ["no-entrypoint"] }
 spl-token-2022 = { version = "4.0.0", path = "../../token/program-2022", features = ["no-entrypoint"] }
