+++ program-test/tests/process_create_associated_token_account.rs
@@ -12,9 +12,8 @@ use {
         signature::Signer,
         transaction::{Transaction, TransactionError},
     },
-    spl_associated_token_account::{
-        get_associated_token_address_with_program_id, instruction::create_associated_token_account,
-    },
+    spl_associated_token_account::instruction::create_associated_token_account,
+    spl_associated_token_account_client::address::get_associated_token_address_with_program_id,
     spl_token_2022::{extension::ExtensionType, state::Account},
 };
