+++ program-test/tests/recover_nested.rs
@@ -14,7 +14,8 @@ use {
         signer::keypair::Keypair,
         transaction::{Transaction, TransactionError},
     },
-    spl_associated_token_account::{get_associated_token_address_with_program_id, instruction},
+    spl_associated_token_account::instruction,
+    spl_associated_token_account_client::address::get_associated_token_address_with_program_id,
     spl_token_2022::{
         extension::{ExtensionType, StateWithExtensionsOwned},
         state::{Account, Mint},
