+++ program-test/tests/spl_token_create.rs
@@ -11,9 +11,8 @@ use {
     solana_program::pubkey::Pubkey,
     solana_program_test::*,
     solana_sdk::{program_pack::Pack, signature::Signer, transaction::Transaction},
-    spl_associated_token_account::{
-        get_associated_token_address, instruction::create_associated_token_account,
-    },
+    spl_associated_token_account::instruction::create_associated_token_account,
+    spl_associated_token_account_client::address::get_associated_token_address,
     spl_token::state::Account,
 };
