+++ program/Cargo.toml
@@ -12,11 +12,11 @@ no-entrypoint = []
 test-sbf = []
 
 [dependencies]
-assert_matches = "1.5.0"
 borsh = "1.5.1"
 num-derive = "0.4"
 num-traits = "0.2"
 solana-program = "2.0.3"
+spl-associated-token-account-client = { version = "1.0.0", path = "../client" }
 spl-token = { version = "6.0", path = "../../token/program", features = [
   "no-entrypoint",
 ] }
