+++ program/src/processor.rs
@@ -5,7 +5,6 @@ use {
         error::AssociatedTokenAccountError,
         instruction::AssociatedTokenAccountInstruction,
         tools::account::{create_pda_account, get_account_len},
-        *,
     },
     borsh::BorshDeserialize,
     solana_program::{
@@ -19,6 +18,7 @@ use {
         system_program,
         sysvar::Sysvar,
     },
+    spl_associated_token_account_client::address::get_associated_token_address_and_bump_seed_internal,
     spl_token_2022::{
         extension::{ExtensionType, StateWithExtensions},
         state::{Account, Mint},
